+++ kame/kame/mip6config/mip6config.h
@@ -4,7 +4,7 @@
  *
  * Author:  Magnus Braathen <magnus.braathen@era.ericsson.se>
  *
- * $Id: mip6config.h,v 1.3 2000/02/09 14:34:02 itojun Exp $
+ * $Id: mip6config.h,v 1.4 2000/02/14 10:04:50 itojun Exp $
  *
  */
 
@@ -12,7 +12,7 @@
 
 struct config_tmpl {
     char *comstring;
-    int (*parse) __P((char *, int));
+    int (*parse) __P((char *, u_long));
 #if defined(__FreeBSD__) && __FreeBSD__ < 3
     int command;
 #else
